requirements.txt
Flask
python-telegram-bot
gspread
oauth2client
pillow
pytesseract
